+++ common/src/main/java/immersive_aircraft/client/render/entity/renderer/AirshipEntityRenderer.java
@@ -9,7 +9,6 @@ import immersive_aircraft.entity.misc.VehicleInventoryDescription;
 import immersive_aircraft.util.Utils;
 import immersive_aircraft.util.obj.Mesh;
 import net.minecraft.block.entity.BannerPattern;
-import net.minecraft.client.MinecraftClient;
 import net.minecraft.client.render.RenderLayer;
 import net.minecraft.client.render.VertexConsumer;
 import net.minecraft.client.render.entity.EntityRenderDispatcher;
@@ -33,7 +32,7 @@ public class AirshipEntityRenderer<T extends AirshipEntity> extends AircraftEnti
             )
             .add(
                     new Object(id, "banners").setRenderConsumer(
-                            (vertexConsumerProvider, entity, matrixStack, light) -> {
+                            (vertexConsumerProvider, entity, matrixStack, light, tickDelta) -> {
                                 List<ItemStack> slots = entity.getSlots(VehicleInventoryDescription.SlotType.BANNER);
                                 int i = 0;
                                 for (ItemStack slot : slots) {
@@ -49,7 +48,7 @@ public class AirshipEntityRenderer<T extends AirshipEntity> extends AircraftEnti
             .add(
                     new Object(id, "sails")
                             .setRenderConsumer(
-                                    (vertexConsumerProvider, entity, matrixStack, light) -> {
+                                    (vertexConsumerProvider, entity, matrixStack, light, tickDelta) -> {
                                         Identifier identifier = getTexture(entity);
                                         VertexConsumer vertexConsumer = vertexConsumerProvider.getBuffer(RenderLayer.getEntityCutoutNoCull(identifier));
 
@@ -67,7 +66,7 @@ public class AirshipEntityRenderer<T extends AirshipEntity> extends AircraftEnti
 
                                         if (entity.isWithinParticleRange() && Config.getInstance().enableAnimatedSails) {
                                             Mesh mesh = getFaces(id, "sails_animated");
-                                            float time = entity.world.getTime() % 24000 + MinecraftClient.getInstance().getTickDelta();
+                                            float time = entity.world.getTime() % 24000 + tickDelta;
                                             renderSailObject(mesh, matrixStack, vertexConsumer, light, time, r, g, b, 1.0f);
                                         } else {
                                             Mesh mesh = getFaces(id, "sails");
@@ -96,7 +95,7 @@ public class AirshipEntityRenderer<T extends AirshipEntity> extends AircraftEnti
                                     }
                             )
                             .setRenderConsumer(
-                                    (vertexConsumerProvider, entity, matrixStack, light) -> {
+                                    (vertexConsumerProvider, entity, matrixStack, light, tickDelta) -> {
                                         Identifier identifier = getTexture(entity);
                                         VertexConsumer vertexConsumer = vertexConsumerProvider.getBuffer(RenderLayer.getEntityCutoutNoCull(identifier));
                                         Mesh mesh = getFaces(id, "propeller");
@@ -111,17 +110,17 @@ public class AirshipEntityRenderer<T extends AirshipEntity> extends AircraftEnti
     }
 
     @Override
-    public Identifier getTexture(T AircraftEntity) {
+    public Identifier getTexture(T entity) {
         return texture;
     }
 
     @Override
-    Model getModel(AircraftEntity entity) {
+    public Model getModel(AircraftEntity entity) {
         return model;
     }
 
     @Override
-    Vec3f getPivot(AircraftEntity entity) {
+    public Vec3f getPivot(AircraftEntity entity) {
         return new Vec3f(0.0f, 0.2f, 0.0f);
     }
 }