+++ common/src/main/java/immersive_aircraft/entity/BiplaneEntity.java
@@ -40,7 +40,7 @@ public class BiplaneEntity extends AirplaneEntity {
     }
 
     @Override
-    float getBaseFuelConsumption() {
+    protected float getBaseFuelConsumption() {
         return 1.25f;
     }