+++ common/src/main/java/immersive_aircraft/entity/EngineAircraft.java
@@ -17,6 +17,7 @@ import net.minecraft.item.Item;
 import net.minecraft.item.ItemStack;
 import net.minecraft.server.network.ServerPlayerEntity;
 import net.minecraft.sound.SoundEvent;
+import net.minecraft.text.Text;
 import net.minecraft.text.TranslatableText;
 import net.minecraft.util.registry.Registry;
 import net.minecraft.world.World;
@@ -28,7 +29,7 @@ import java.util.Map;
  * Simulated engine behavior
  */
 public abstract class EngineAircraft extends AircraftEntity {
-    static final TrackedData<Float> ENGINE = DataTracker.registerData(EngineAircraft.class, TrackedDataHandlerRegistry.FLOAT);
+	protected static final TrackedData<Float> ENGINE = DataTracker.registerData(EngineAircraft.class, TrackedDataHandlerRegistry.FLOAT);
     private static final TrackedData<Float> UTILIZATION = DataTracker.registerData(EngineAircraft.class, TrackedDataHandlerRegistry.FLOAT);
     private static final TrackedData<Boolean> LOW_ON_FUEL = DataTracker.registerData(EngineAircraft.class, TrackedDataHandlerRegistry.BOOLEAN);
 
@@ -37,7 +38,7 @@ public abstract class EngineAircraft extends AircraftEntity {
     public float engineSpinUpStrength = 0.0f;
     public float engineSound = 0.0f;
 
-    enum FuelState {
+    protected enum FuelState {
         NEVER,
         EMPTY,
         FUELED,
@@ -66,27 +67,27 @@ public abstract class EngineAircraft extends AircraftEntity {
         fuel = new int[getInventoryDescription().getSlots(VehicleInventoryDescription.SlotType.BOILER).size()];
     }
 
-    SoundEvent getEngineStartSound() {
+    protected SoundEvent getEngineStartSound() {
         return Sounds.ENGINE_START.get();
     }
 
-    SoundEvent getEngineSound() {
+    protected SoundEvent getEngineSound() {
         return Sounds.PROPELLER.get();
     }
 
-    float getEnginePitch() {
+    protected float getEnginePitch() {
         return 1.0f;
     }
 
-    float getStabilizer() {
+    protected float getStabilizer() {
         return 0.0f;
     }
 
-    float getBaseFuelConsumption() {
+    protected float getBaseFuelConsumption() {
         return 0.75f;
     }
 
-    float getEngineReactionSpeed() {
+    protected float getEngineReactionSpeed() {
         return 20.0f;
     }
 
@@ -190,7 +191,7 @@ public abstract class EngineAircraft extends AircraftEntity {
         }
     }
 
-    String getFuelType() {
+    protected String getFuelType() {
         return "fuel";
     }
 
@@ -224,7 +225,7 @@ public abstract class EngineAircraft extends AircraftEntity {
     }
 
     @Override
-    void updateController() {
+    protected void updateController() {
         // left-right
         setYaw(getYaw() - getProperties().getYawSpeed() * pressingInterpolatedX.getSmooth());
 
@@ -236,7 +237,7 @@ public abstract class EngineAircraft extends AircraftEntity {
     }
 
     @Override
-    void updateVelocity() {
+    protected void updateVelocity() {
         super.updateVelocity();
 
         // landing