+++ common/src/main/java/immersive_aircraft/entity/GyrodyneEntity.java
@@ -10,6 +10,7 @@ import net.minecraft.entity.player.PlayerEntity;
 import net.minecraft.item.Item;
 import net.minecraft.server.network.ServerPlayerEntity;
 import net.minecraft.sound.SoundEvent;
+import net.minecraft.text.Text;
 import net.minecraft.text.TranslatableText;
 import net.minecraft.util.math.MathHelper;
 import net.minecraft.util.math.Vec3d;
@@ -55,16 +56,16 @@ public class GyrodyneEntity extends Rotorcraft {
         super(entityType, world);
     }
 
-    SoundEvent getEngineStartSound() {
+    protected SoundEvent getEngineStartSound() {
         return Sounds.WOOSH.get();
     }
 
-    SoundEvent getEngineSound() {
+    protected SoundEvent getEngineSound() {
         return Sounds.WOOSH.get();
     }
 
     @Override
-    float getStabilizer() {
+    protected float getStabilizer() {
         return 0.3f;
     }
 
@@ -74,17 +75,17 @@ public class GyrodyneEntity extends Rotorcraft {
     }
 
     @Override
-    float getGroundVelocityDecay() {
+    protected float getGroundVelocityDecay() {
         return falloffGroundVelocityDecay(0.8f);
     }
 
     @Override
-    float getHorizontalVelocityDelay() {
+    protected float getHorizontalVelocityDelay() {
         return 0.925f;
     }
 
     @Override
-    float getVerticalVelocityDelay() {
+    protected float getVerticalVelocityDelay() {
         return 0.9f;
     }
 
@@ -120,7 +121,7 @@ public class GyrodyneEntity extends Rotorcraft {
     }
 
     @Override
-    String getFuelType() {
+    protected String getFuelType() {
         return "fat";
     }
 
@@ -130,7 +131,7 @@ public class GyrodyneEntity extends Rotorcraft {
     }
 
     @Override
-    void updateController() {
+    protected void updateController() {
         super.updateController();
 
         // launch that engine