+++ common/src/main/java/immersive_aircraft/entity/InventoryVehicleEntity.java
@@ -8,6 +8,7 @@ import immersive_aircraft.item.upgrade.AircraftStat;
 import immersive_aircraft.network.s2c.OpenGuiRequest;
 import immersive_aircraft.screen.VehicleScreenHandler;
 import net.minecraft.enchantment.EnchantmentHelper;
+import net.minecraft.entity.Entity;
 import net.minecraft.entity.EntityType;
 import net.minecraft.entity.player.PlayerEntity;
 import net.minecraft.entity.player.PlayerInventory;
@@ -103,7 +104,7 @@ public abstract class InventoryVehicleEntity extends VehicleEntity implements In
         return new VehicleScreenHandler(i, playerInventory, this);
     }
 
-    int syncId;
+    protected int syncId;
 
     public void openInventory(ServerPlayerEntity player) {
         syncId = (syncId + 1) % 100 + 100;
@@ -111,14 +112,21 @@ public abstract class InventoryVehicleEntity extends VehicleEntity implements In
         if (screenHandler != null) {
             NetworkHandler.sendToPlayer(new OpenGuiRequest(this, screenHandler.syncId), player);
             player.currentScreenHandler = screenHandler;
-            player.currentScreenHandler.addListener(player);
+             player.currentScreenHandler.addListener(player);
         }
     }
 
     @Override
     public ActionResult interact(PlayerEntity player, Hand hand) {
         if (!player.world.isClient && player.shouldCancelInteraction()) {
-            openInventory((ServerPlayerEntity)player);
+            Entity primaryPassenger = getFirstPassenger();
+            if (primaryPassenger != null) {
+                // Kick out the first passenger
+                primaryPassenger.stopRiding();
+            } else {
+                // Open inventory instead
+                openInventory((ServerPlayerEntity) player);
+            }
             return ActionResult.CONSUME;
         }
         return super.interact(player, hand);