+++ common/src/main/java/immersive_aircraft/network/c2s/CommandMessage.java
@@ -4,10 +4,11 @@ import immersive_aircraft.cobalt.network.Message;
 import immersive_aircraft.entity.InventoryVehicleEntity;
 import immersive_aircraft.entity.VehicleEntity;
 import net.minecraft.entity.player.PlayerEntity;
+import net.minecraft.network.PacketByteBuf;
 import net.minecraft.server.network.ServerPlayerEntity;
 import net.minecraft.util.math.Vec3d;
 
-public class CommandMessage implements Message {
+public class CommandMessage extends Message {
     private final Key key;
     private final double fx;
     private final double fy;
@@ -18,6 +19,21 @@ public class CommandMessage implements Message {
         this.fx = velocity.x;
         this.fy = velocity.y;
         this.fz = velocity.z;
+    }
+
+    public CommandMessage(PacketByteBuf b) {
+        key = Key.values()[b.readInt()];
+        fx = b.readDouble();
+        fy = b.readDouble();
+        fz = b.readDouble();
+    }
+
+    @Override
+    public void encode(PacketByteBuf b) {
+        b.writeInt(key.ordinal());
+        b.writeDouble(fx);
+        b.writeDouble(fy);
+        b.writeDouble(fz);
     }
 
     @Override