+++ common/src/main/java/immersive_aircraft/network/c2s/EnginePowerMessage.java
@@ -3,12 +3,22 @@ package immersive_aircraft.network.c2s;
 import immersive_aircraft.cobalt.network.Message;
 import immersive_aircraft.entity.EngineAircraft;
 import net.minecraft.entity.player.PlayerEntity;
+import net.minecraft.network.PacketByteBuf;
 
-public class EnginePowerMessage implements Message {
+public class EnginePowerMessage extends Message {
     private final float engineTarget;
 
     public EnginePowerMessage(float engineTarget) {
         this.engineTarget = engineTarget;
+    }
+
+    public EnginePowerMessage(PacketByteBuf b) {
+        engineTarget = b.readFloat();
+    }
+
+    @Override
+    public void encode(PacketByteBuf b) {
+        b.writeFloat(engineTarget);
     }
 
     @Override